requirements.txt
langchain-google-genai
langchain 
streamlit 
faiss-cpu
unstructured 
PyMuPDF 
opencv-python 
pillow 
python-dotenv
easyocr
crawl4ai 
beautifulsoup4 
requests
sentence-transformers 
torch
langchain-community
pdfminer.six
pi-heif
unstructured-inference
pdf2image
numpy
